Update patch set 3
Patch Set 3:

> Hi Khen, We made changes to a few APIs (Get, Put, List and Delete)
 > as per your suggestion. The changes were in etcdclient.go,
 > client.go, consulclient.go, backend.go, transaction.go (rw_core).
 > There will be changes in ro_core and adapter as well. Moreover, we
 > are not passing the context from the GRPC API layer; instead we
 > just created at backend.go.
 > We wanted to check with you about an alternative wherein we add the
 > connTimeout as a variable in the EtcdClient Struct. We can fix a
 > default timeout which can be overriden by the user of the object.

Thanks Thomas.   I think adding a context to the APIs is the right approach.   Changing the API is a major change, meaning the library major version will need changing as well.  The dependent components can migrate to this new library version at their convenience.   You may want to bring this discussion to the TST as there are other changes that need to be done in the library and coordination of these will be required.

Patch-set: 3
Label: Code-Review=0