Update patch set 3
Patch Set 3:

> (1 comment)

Got it. Can I address this as part of the next story I am working on related to deletion of tech-profile. As Matt was saying, we should make the instance-control values as enums which will need a change in techprofile.proto; too many cross dependencies. Let's close this review and I will address this as part of next review. Is that fine?

Patch-set: 3