Update patch set 7
Patch Set 7: Code-Review-1

(2 comments)

Downgrading to -1 as we don't need the InvalidLogLevel at this point. As a general question now that we are expecting the StringToInt to return an error.

As a more general question should there be a type for LogLevel as opposed to just int.

type LogLevel int
const (
    DebugLevel = LogLevel(iota)
    InfoLevel
...

Patch-set: 7
Label: Code-Review=-1