Update patch set 7
Patch Set 7:

> (1 comment)

I think we should set the value to 60 secs to cater for load testing.  Once adapters changes their code then the timeout won't be used.   The other option is to remove the timeout completely  and force adapter code change to specify a timeout with a context.   Note sure how much changes that would require in the adapters though.

Patch-set: 7