[VOL-2831] WIP - Support multiple topic per deviceType
**** Expect jenkins to fail as proto dependency not released ***

This is the initial commit for multiple adapter support.  While
there are a number of files in this commit the main file is the
endpoint_manager.go.   The changes in the mocks directory is only
to work around a package dependency in tests. The go.mod replaces
the protos to a local copy (the path needs to be changed to your
specific directory)

The endpoint_manager.go provides two APIs that will be needed
by components communicating over kafka:
 - GetEndPoint() : to be called before sending a request to kafka
 - IsDeviceOwnedByService(): used for reconciliation

Change-Id: I35896451f1d8fd40226b18499cbe13868f5844bc