Update patch set 3
Patch Set 3: Code-Review-1

In general, there is no point passing ctx down to everything.  Most functions don't need ctx, so it's just an extra argument to keep track of.  (In some places, it should be passed, but a quick "pass it everywhere" won't fix underlying request cancellation problems.)  

Is this being done to facilitate context-based logging?  Perhaps only pass ctx where it's required to do logging?

I'm not totally convinced that this is the right solution, and am hesitant because of all the code churn this will cause.

If context-based logging is the way forward, it should be done with a logging cleanup, to reduce the amount of redundant dubug logging, and thus reduce how many places a ctx needs to be passed.

Patch-set: 3
Reviewer: Gerrit User 1000107 <1000107@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=-1