Update patch set 3
Patch Set 3:

Yes, you are right. In this patch, the context is being passed for use in logging method calls so as to achieve correlation among the logs generated by a flow of execution. Unfortunately, this is only standard approach possible in golang due to the absence of Thread-local Storage. Additionally, the tracing span would also be carried through the context propagation. Please refer to follow doc on the solution approach (still being detailed out).
https://docs.google.com/document/d/1GiElGlfxvfBR18k7Xip-JGHZzgCq779c/

I do agree that it is not required to pass context into each method call as it is not used in each method. But, it is needed in close to 60% of the methods. So, we felt it is more appropriate to put it as first argument of each function to ensure consistency. This will ensure new methods added also follow the convention and we can even introduce a SCA check later in build to check for ctx as first argument to enforce it easily.

Yes, this change is going to affect the whole codebase but it is needed to implement Log Correlation and Tracing feature into Voltha components. We are approaching it one step at a time, avoiding any build issues. Please check the other related patchset for VOL-3071 (https://gerrit.opencord.org/#/c/18996/) where a new logger has been introduced with ctx as first argument but previous logger will still continue to work while we gradually migrate.

Patch-set: 3
Reviewer: Gerrit User 1000526 <1000526@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0