Update patch set 3
Patch Set 3:

> Yes, you are right. In this patch, the context is being passed for
 > use in logging method calls so as to achieve correlation among the
 > logs generated by a flow of execution. Unfortunately, this is only
 > standard approach possible in golang due to the absence of
 > Thread-local Storage. Additionally, the tracing span would also be
 > carried through the context propagation. Please refer to follow doc
 > on the solution approach (still being detailed out).
 > https://docs.google.com/document/d/1GiElGlfxvfBR18k7Xip-JGHZzgCq779c/
 > 
 > I do agree that it is not required to pass context into each method
 > call as it is not used in each method. But, it is needed in close
 > to 60% of the methods. So, we felt it is more appropriate to put it
 > as first argument of each function to ensure consistency. This will
 > ensure new methods added also follow the convention and we can even
 > introduce a SCA check later in build to check for ctx as first
 > argument to enforce it easily.
 > 
 > Yes, this change is going to affect the whole codebase but it is
 > needed to implement Log Correlation and Tracing feature into Voltha
 > components. We are approaching it one step at a time, avoiding any
 > build issues. Please check the other related patchset for VOL-3071
 > (https://gerrit.opencord.org/#/c/18996/) where a new logger has
 > been introduced with ctx as first argument but previous logger will
 > still continue to work while we gradually migrate.

I agree in general, that if context is needed for logging it should be passed.

However, I disagree strongly that ctx should be passed to _every_ call.  It should only be passed where it is needed.  If ~60% of calls require ctx to be passed to enable logging, then in ~40% of cases it's completely unnecessary.

In the patch that handles logging updates, ctx can be added to only the functions that need it by following a simple process.  (Can just find&replace log statements, then add ctx to any functions that no longer compile)

I've added an item to discuss this specific issues at the next TST, (let me know if this discussion should happen elsewhere).

Patch-set: 3