Update patch set 4
Patch Set 4:

I see a number of places where a background context is created and add to the logger call.  These obviously, won't have the metadata required for jaeger.   Do we want to do all these changes in this update?   Can we only address the log changes and methods where we already have a context as a parameter since only those could be traceable. Then create separate Jiras to add context to methods and their dependencies.  We could have a separate Jira per package.   Would that be easier or I am completely out of context?

Patch-set: 4
Label: Code-Review=0