Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (2 comments)
> 
> There are old contexts being used when receiving messages from kafka.  
> 
> Is the plan to make changes to rw-core and openolt adapter, test them together and then merge the entire set together?


@khen..For now, we have done the changes for adding the ctx in the logger and passing the ctx across the methods. These changes have been done in voltha-lib-go(VOL-3069), openolt-adapter(VOL-3069), voltha-go(VOL-3187) and ofagent-go(VOL-3187) 
Further, on top of this, in another jira, we would take the opentracing span related changes where we will be taking the span from old ctx and a new ctx would be created and passed further. We will be handing the go routines also.

Previous discussions on this jira: https://gerrit.opencord.org/c/voltha-lib-go/+/19201

Patch-set: 1