Update patch set 3
Patch Set 3: Code-Review-1

I'm not sure if this would solve the problem described in VOL-3120. The need is that first check if there is a TP data inserted to the etcd specified by the environment variable. And if there is no TP data in that path then check the default TP path location. So you'll need to access and find where is the TP data located.

Patch-set: 3
Reviewer: Gerrit User 1000404 <1000404@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=-1