Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> this needs rebase, but the approach is ok to me.

Hi guys,
Let me rephrase the issue.
The TP data is inserted by the NEM to etcd. In the case of multiple Voltha stacks, we need to provide a unique KV_STORE_PATH to each stack. If the NEM has already knows all of the unique data paths in advance, it can insert the TP in every location. If that's not possible then it would insert it to a default location. Therefore, a Voltha stack need to know where to read the TP data. I'm not sure when the TP data is inserted and how it's used but this was the original problem. 
This patch would have no effect in case of multiple Voltha stacks since we have to define a unique kv path to each stack and it would never fall back to the default path. 
The problem can also be solved by defining a special path for the TP data. 
Any thoughts?

Patch-set: 4
Reviewer: Gerrit User 1000404 <1000404@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>