Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 3:
> > 
> > this needs rebase, but the approach is ok to me.
> 
> Hi guys,
> Let me rephrase the issue.
> The TP data is inserted by the NEM to etcd. In the case of multiple Voltha stacks, we need to provide a unique KV_STORE_PATH to each stack. If the NEM has already knows all of the unique data paths in advance, it can insert the TP in every location. If that's not possible then it would insert it to a default location. Therefore, a Voltha stack need to know where to read the TP data. I'm not sure when the TP data is inserted and how it's used but this was the original problem. 
> This patch would have no effect in case of multiple Voltha stacks since we have to define a unique kv path to each stack and it would never fall back to the default path. 
> The problem can also be solved by defining a special path for the TP data. 
> Any thoughts?

I believe the most robust solution for this issue would be to push TP json into Voltha through a gRPC API which unfortunately was never built.

Actually, TP would be pushed by NEM on the fly (as and when they are configured by XOS TOSCA layer) and not only at SEBA startup. So, putting this logic at Voltha startup to finalize the initial kvStoreDataPathPrefix is not a practical solution. Initially, nothing would be in the TP path.

We would need to handle this challenge when TP is read from etcd against the defined Technology + IO combination. We should store both default and Stack specific path in TP manager. One possible solution would be to perform following steps everytime a TP instance is accessed:
1. Check if TP instance is present under Stack specific path in etcd. If found, load and use it
2. Else, check if TP instance is present under Default TP path in etcd. If found, load and use it. Also, save it under Stack specific path.
3. If not found, fallback to default TP ID 64 and try to load it from Stack specific path in etcd. If found, load and use it.
4. Else, create a default TP instance with TP ID 64 and save it into etcd.

Step 3 and 4 are already implemented.

As there would be just a handful of TP instances per operator, this approach should not lead to many additional steps.

Patch-set: 4
Reviewer: Gerrit User 1000526 <1000526@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>