Update patch set 4
Patch Set 4:

Sorry for chiming in late here, wasn't part of the review or discussion earlier.

After looking at all the discussions on this review - Just to add my two cents, NEM should be aware of the TP path where it is going to upload the TPs. It does not matter when it adds it, all that matters is when subscriber is provisioned with a TP-ID and adapter makes a query to TP Module, the TP should be present at the given TP path. If this is the only requirement, then the current patch should solve the problem. From what I understand, deployment charts are going to set different 'TP path prefix' environment variable for different stacks, which sounds fine.

Patch-set: 4
CC: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>