Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Interesting, can you elaborate a little more on why this provides a fix ?

Checked the code for sarama, there is a check for adding timestamp which is 
ps.parent.conf.Version.IsAtLeast(V0_10_0_0) and if we donot provide any version, default it was taking to 0.8.2 and If we see the createClusterAdmin function we were adding version sarama.V1_0_0_0, so added same version wherever we are using NewConfig. Also a couple of bugs are there on sarama where timestamp was not getting added and the solution suggested was to check the version.

Patch-set: 1