Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Will this patch solve the issue described in https://jira.opencord.org/browse/VOL-3908?

Yes, with current images and logs given on the jira, kafka timestamp is coming as -1 epoch time which is equal to '1969-12-31 23:59:59.999000' . After building core and adapters with this image of lib, proper epoch time is coming which when converted to date time is proper. You can also test by taking this version of lib and building core, and adapters.

Patch-set: 1