Update patch set 19
Patch Set 19:

> Patch Set 19:
> 
> (1 comment)
> 
> Why was this change needed? I do not see any details on the jira.

Girish, this change allows the sending of kafka event to be asynchronous.  In other words, the caller will publish the event to the proxy and leave.  The proxy takes care of sending the event onto kafka in the order in which the events were received. There are improvements needed in this change, as mentioned by David, notably to have a threshold  around the number of go routines that gets created.

Patch-set: 19
Reviewer: Gerrit User 1000056 <1000056@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>