[VOL-4194] Add an Etcd Client Pool
This commit adds an Etcd client pool with a configurable
capacity as well as maximum concurrent requests per client.  While
tests were done locally, it was not done with all components
due to dependencies that need to be merged first.   Once the
dependencies are merged then this change can be fully tested.

Change-Id: I7e8c8953bd3871056a721de68990e3d85df8b688